updated file versions. duration_prophylaxis now can be specified as list. More flexible checks on inputs that replicate scalar deme properties over demes by default
