+++ src/Haplo_node.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "Sampler_v3.h"
+#include "Sampler_v4.h"
 
 #include <map>
 #include <vector>
